+++ M5StackCommon/TouchEventCategory.cs
@@ -1,9 +1,16 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
+#if TOUGH || M5CORE2
 using System;
+#endif
 
+#if TOUGH
+namespace nanoFramework.Tough
+#elif M5CORE2
 namespace nanoFramework.M5Core2
+#endif
+#if TOUGH || M5CORE2
 {
     /// <summary>
     /// Sub event touch catgory
@@ -33,3 +40,4 @@ namespace nanoFramework.M5Core2
         LiftUp = 0b0010_0000,
     }
 }
+#endif
