+++ nanoFramework.Tough/Tough.cs
@@ -1,0 +1,12 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+namespace nanoFramework.M5Stack
+{
+    /// <summary>
+    /// Tough board
+    /// </summary>
+    public static partial class Tough
+    {
+    }
+}
