+++ nanoFramework.Tough/packages.config
@@ -1,0 +1,28 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="nanoFramework.CoreLibrary" version="1.12.0" targetFramework="netnanoframework10" />
+  <package id="nanoFramework.Graphics" version="1.0.2" targetFramework="netnanoframework10" />
+  <package id="nanoFramework.Hardware.Esp32" version="1.3.6" targetFramework="netnanoframework10" />
+  <package id="nanoFramework.Iot.Device.Axp192" version="1.1.67.25390" targetFramework="netnano1.0" />
+  <package id="nanoFramework.Iot.Device.Chs6540" version="1.0.16" targetFramework="netnanoframework10" />
+  <package id="nanoFramework.Iot.Device.Common.NumberHelper" version="1.1.1" targetFramework="netnanoframework10" />
+  <package id="nanoFramework.Iot.Device.Rtc" version="1.1.61.27737" targetFramework="netnano1.0" />
+  <package id="nanoFramework.ResourceManager" version="1.1.4" targetFramework="netnanoframework10" />
+  <package id="nanoFramework.Runtime.Events" version="1.10.0" targetFramework="netnanoframework10" />
+  <package id="nanoFramework.Runtime.Native" version="1.5.4" targetFramework="netnanoframework10" />
+  <package id="nanoFramework.System.Collections" version="1.4.0" targetFramework="netnanoframework10" />
+  <package id="nanoFramework.System.Device.Adc" version="1.0.2" targetFramework="netnanoframework10" />
+  <package id="nanoFramework.System.Device.Dac" version="1.4.3" targetFramework="netnanoframework10" />
+  <package id="nanoFramework.System.Device.Gpio" version="1.0.4" targetFramework="netnanoframework10" />
+  <package id="nanoFramework.System.Device.I2c" version="1.0.3" targetFramework="netnanoframework10" />
+  <package id="nanoFramework.System.Device.Spi" version="1.2.1" targetFramework="netnanoframework10" />
+  <package id="nanoFramework.System.IO.Ports" version="1.0.7.1" targetFramework="netnano1.0" />
+  <package id="nanoFramework.System.IO.Streams" version="1.0.0" targetFramework="netnanoframework10" />
+  <package id="nanoFramework.System.Text" version="1.1.3" targetFramework="netnanoframework10" />
+  <package id="nanoFramework.System.Threading" version="1.0.4" targetFramework="netnanoframework10" />
+  <package id="Nerdbank.GitVersioning" version="3.5.103" targetFramework="netnanoframework10" developmentDependency="true" />
+  <package id="UnitsNet.nanoFramework.ElectricCurrent" version="4.131.0" targetFramework="netnano1.0" />
+  <package id="UnitsNet.nanoFramework.ElectricPotential" version="4.131.0" targetFramework="netnano1.0" />
+  <package id="UnitsNet.nanoFramework.Power" version="4.131.0" targetFramework="netnano1.0" />
+  <package id="UnitsNet.nanoFramework.Temperature" version="4.131.0" targetFramework="netnano1.0" />
+</packages>
